fix bugs

+++ src/shaders/pbr.frag
@@ -10,7 +10,7 @@
 in highp vec3 position;
 in highp vec3 normal;
 #if defined(TEXTURED)
-in mediump vec2 texCoord;
+in highp vec2 texCoord;
 #endif
 #if defined(NORMAL_TEXTURE) && defined(PRECOMPUTED_TANGENT)
 in highp vec3 tangent;
